Navtabs component JSS styles in separate file navtabsStyle.js
